-- renamed the packages -- added App.java entrance to the programme
