--- src/main/java/ConsoleControl/CommandList.java
@@ -1,13 +1,0 @@
-package ConsoleControl;
-
-public enum CommandList {
-  ONLINE_BOARD,
-  SHOW_FLIGHT_INFO,
-  SEARCH_AND_BOOK_FLIGHT,
-  CANCEL_BOOKING,
-  MY_BOOKINGS,
-  END_SESSION,
-  HELP,
-  EXIT
-
-}
