--- src/main/java/ConsoleControl/Printer/FancyString/FancyString.java
@@ -1,5 +1,0 @@
-package ConsoleControl.Printer.FancyString;
-
-public interface FancyString {
-  String fancyString();
-}
