--- src/main/java/ConsoleControl/Printer/Printer.java
@@ -1,13 +1,0 @@
-package ConsoleControl.Printer;
-
-import ConsoleControl.Printer.FancyString.FancyString;
-
-import java.util.List;
-
-public interface Printer {
-  static void print(List<FancyString> list) {
-    for (FancyString el : list) {
-      System.out.println(el.fancyString());
-    }
-  }
-}
