+++ src/main/java/consoleControl/CommandList.java
@@ -1,0 +1,13 @@
+package consoleControl;
+
+public enum CommandList {
+  ONLINE_BOARD,
+  SHOW_FLIGHT_INFO,
+  SEARCH_AND_BOOK_FLIGHT,
+  CANCEL_BOOKING,
+  MY_BOOKINGS,
+  END_SESSION,
+  HELP,
+  EXIT
+
+}
