+++ src/main/java/consoleControl/Printer/FancyString/FancyString.java
@@ -1,0 +1,5 @@
+package consoleControl.Printer.FancyString;
+
+public interface FancyString {
+  String fancyString();
+}
