+++ src/main/java/consoleControl/Printer/Printer.java
@@ -1,0 +1,13 @@
+package consoleControl.Printer;
+
+import consoleControl.Printer.FancyString.FancyString;
+
+import java.util.List;
+
+public interface Printer {
+  static void print(List<FancyString> list) {
+    for (FancyString el : list) {
+      System.out.println(el.fancyString());
+    }
+  }
+}
