+++ src/main/java/dao/DAO.java
@@ -1,0 +1,11 @@
+package dao;
+
+public interface DAO<T> {
+  void add(T data);
+
+  void remove(String id);
+
+  T get(String id);
+
+  void update(T data);
+}
