--- src/test/java/ConsoleControl/CommandParserTest.java
@@ -1,5 +1,0 @@
-package ConsoleControl;
-
-public class CommandParserTest {
-
-}
