+++ src/test/java/consoleControl/CommandParserTest.java
@@ -1,0 +1,5 @@
+package consoleControl;
+
+public class CommandParserTest {
+
+}
